moved TokenClaim.java to 'security' package
